requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.16.2
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
asgiref==3.8.1
    # via opentelemetry-instrumentation-asgi
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via posthog
bcrypt==4.3.0
    # via chromadb
blinker==1.9.0
    # via flask
build==1.2.2.post1
    # via chromadb
cachetools==5.5.2
    # via google-auth
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
charset-normalizer==3.4.2
    # via requests
chromadb==1.0.12
    # via -r requirements.in
click==8.2.1
    # via
    #   flask
    #   typer
    #   uvicorn
coloredlogs==15.0.1
    # via onnxruntime
deprecated==1.2.18
    # via limits
distro==1.9.0
    # via
    #   openai
    #   posthog
durationpy==0.10
    # via kubernetes
fastapi==0.115.9
    # via chromadb
filelock==3.18.0
    # via huggingface-hub
flask==3.1.1
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-limiter
    #   flask-pydantic
    #   prometheus-flask-exporter
flask-cors==6.0.1
    # via -r requirements.in
flask-limiter==3.12
    # via -r requirements.in
flask-pydantic==0.13.1
    # via -r requirements.in
flatbuffers==25.2.10
    # via onnxruntime
fsspec==2025.5.1
    # via huggingface-hub
google-auth==2.40.3
    # via kubernetes
googleapis-common-protos==1.70.0
    # via opentelemetry-exporter-otlp-proto-grpc
grpcio==1.73.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.1.4
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   chromadb
    #   openai
huggingface-hub==0.33.0
    # via tokenizers
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
importlib-resources==6.5.2
    # via chromadb
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jiter==0.10.0
    # via openai
jsonschema==4.24.0
    # via chromadb
jsonschema-specifications==2025.4.1
    # via jsonschema
kubernetes==33.1.0
    # via chromadb
limits==5.4.0
    # via flask-limiter
loguru==0.7.3
    # via -r requirements.in
mako==1.3.10
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   mako
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
mmh3==5.1.0
    # via chromadb
mpmath==1.3.0
    # via sympy
numpy==2.3.0
    # via
    #   chromadb
    #   onnxruntime
oauthlib==3.2.2
    # via
    #   kubernetes
    #   requests-oauthlib
onnxruntime==1.22.0
    # via chromadb
openai==1.88.0
    # via -r requirements.in
opentelemetry-api==1.34.1
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.34.1
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-grpc==1.34.1
    # via chromadb
opentelemetry-instrumentation==0.55b1
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi==0.55b1
    # via opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-fastapi==0.55b1
    # via chromadb
opentelemetry-proto==1.34.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.34.1
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions==0.55b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-sdk
opentelemetry-util-http==0.55b1
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
ordered-set==4.1.0
    # via flask-limiter
orjson==3.10.18
    # via chromadb
overrides==7.7.0
    # via chromadb
packaging==25.0
    # via
    #   build
    #   huggingface-hub
    #   limits
    #   onnxruntime
    #   opentelemetry-instrumentation
posthog==5.0.0
    # via chromadb
prometheus-client==0.22.1
    # via prometheus-flask-exporter
prometheus-flask-exporter==0.23.2
    # via -r requirements.in
protobuf==5.29.5
    # via
    #   googleapis-common-protos
    #   onnxruntime
    #   opentelemetry-proto
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   chromadb
    #   fastapi
    #   flask-pydantic
    #   openai
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.1
    # via rich
pypika==0.48.9
    # via chromadb
pyproject-hooks==1.2.0
    # via build
python-dateutil==2.9.0.post0
    # via
    #   kubernetes
    #   posthog
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.2
    # via
    #   chromadb
    #   huggingface-hub
    #   kubernetes
    #   uvicorn
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   huggingface-hub
    #   kubernetes
    #   posthog
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via kubernetes
rich==13.9.4
    # via
    #   chromadb
    #   flask-limiter
    #   typer
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   kubernetes
    #   posthog
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via
    #   -r requirements.in
    #   alembic
starlette==0.45.3
    # via fastapi
sympy==1.14.0
    # via onnxruntime
tenacity==9.1.2
    # via chromadb
tokenizers==0.21.1
    # via chromadb
tqdm==4.67.1
    # via
    #   chromadb
    #   huggingface-hub
    #   openai
typer==0.16.0
    # via chromadb
typing-extensions==4.14.0
    # via
    #   alembic
    #   anyio
    #   chromadb
    #   fastapi
    #   huggingface-hub
    #   limits
    #   openai
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.4.0
    # via
    #   kubernetes
    #   requests
uvicorn[standard]==0.34.3
    # via chromadb
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websocket-client==1.8.0
    # via kubernetes
websockets==15.0.1
    # via uvicorn
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
wrapt==1.17.2
    # via
    #   deprecated
    #   opentelemetry-instrumentation
zipp==3.23.0
    # via importlib-metadata
